// Frontend/src/app/models/informacion-economica.ts
export interface InformacionEconomica {
  objSocial: string;
  codCiiu: string;
  docConst: string;
  nroMatricula: string;
  capital: string;
  servicio: string;
  fecha: Date;
  certIso: string;
  fechaInicio: Date;
  implementado: string;
  enProceso: boolean;
  clasificacion: string;
  opcionesContables: string;
  otroContable: string;
  condicionesPago: string;
  otroCondicionPago: string;
}